+++ ReleaseNotes.md
@@ -1,8 +1,6 @@
 0.2.0 - 2021-06-05
 ==================
-Support OAuth 2.0. For details of Slack OAuth 2.0 v2 see
-- https://api.slack.com/authentication/oauth-v2
-- https://api.slack.com/methods/oauth.v2.access
+1. New stack `cdk deploy K-CDK-SlackApp-OAuth` to support OAuth 2.0 flow.
 
 0.1.0 - 2021-05-15
 ==================